requirements.txt
# Core Dependencies
streamlit>=1.28.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-google-genai>=0.0.8
google-generativeai>=0.3.0
chromadb>=0.4.20
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0

# Data Processing
pypdf>=3.17.0
python-docx>=1.1.0
openpyxl>=3.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
sentence-transformers>=2.2.0

# Utilities  
pydantic>=2.5.0
click>=8.1.0
tqdm>=4.66.0
loguru>=0.7.0
tenacity>=8.2.0
faiss-cpu>=1.7.4

# Development Dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.9.0
isort>=5.12.0
flake8>=6.1.0
